feat: error page and drafts moved to api
